split out evaluation and initialization files
